+++ interfaces/datasource.d.ts
@@ -1,7 +1,0 @@
-import '@frontastic/extension-types';
-
-declare module '@frontastic/extension-types' {
-  interface DataSourceConfiguration {
-    streamId?: string;
-  }
-}