+++ interfaces/frontastic-augmentation.d.ts
@@ -8,25 +8,21 @@
 
 import '@frontastic/extension-types';
 import { Token } from '@Types/Token';
-import { Account } from '@Types/account';
+import { ProductListDataSourceConfiguration } from '@Commerce-commercetools/interfaces/DataSource';
 
-// module to add sessionData and query types instead of any
 declare module '@frontastic/extension-types' {
-  export interface Request<T = never> {
-    query: T;
+  interface Request {
     sessionData: SessionData;
   }
+  export interface DataSourceConfiguration {
+    configuration: ProductListDataSourceConfiguration;
+  }
 }
+
 interface SessionData {
   checkoutSessionToken?: Record<string, Token>;
   accountId?: string;
+  accountGroupId?: string;
   wishlistId?: string;
   cartId?: string;
-  account: Account;
-}
-export interface RawFacetData {
-  min?: string | number;
-  max?: string | number;
-  terms?: string[] | Record<string, string>;
-  boolean?: string | boolean;
 }