+++ mappers/CartMapper.ts
@@ -427,9 +427,10 @@ export class CartMapper {
         description:
           LocalizedValue.getLocalizedValue(locale, defaultLocale, commercetoolsDiscountCode.description) || undefined,
         discounts: commercetoolsDiscountCode.cartDiscounts.map((commercetoolsCartDiscount) => {
-          return {
-            cartDiscountId: commercetoolsCartDiscount.id,
-          };
+          if (this.isCartDiscountReferenceExpanded(commercetoolsCartDiscount)) {
+            return this.commercetoolsCartDiscountToCartDiscount(commercetoolsCartDiscount.obj, locale, defaultLocale);
+          }
+          return undefined;
         }),
       };
     }
@@ -534,6 +535,17 @@ export class CartMapper {
       case 'totalPrice':
         return {
           type: 'totalPrice',
+        };
+      case 'multiBuyLineItems':
+        return {
+          type: 'multiBuyLineItems',
+          predicate: commercetoolsCartDiscountTarget.predicate,
+          selectionMode: this.commercetoolsSelectionModeToCartDiscountSelectionMode(
+            commercetoolsCartDiscountTarget.selectionMode,
+          ),
+          triggerQuantity: commercetoolsCartDiscountTarget.triggerQuantity,
+          discountedQuantity: commercetoolsCartDiscountTarget.discountedQuantity,
+          maxOccurrence: commercetoolsCartDiscountTarget.maxOccurrence,
         };
       default:
         return undefined;